requirements.txt
faster_whisper
pysubs2
tqdm
whisper-ctranslate2         # https://github.com/Softcatala/whisper-ctranslate2
srtranslator                # translator
opencc-python-reimplemented # chinese conversion
yt-dlp                     # yt download
loguru                      # logger
